requirements.txt
matplotlib==3.2.1
numpy==1.18.4
mysql_connector_repackaged==0.3.1
